Rename EMS.Judge.Application.Core to Common
