+++ src/Judge/EMS.Judge.Application/Queries/HorseQueries.cs
@@ -1,6 +1,6 @@
 ﻿using Core.Domain.State;
 using Core.Domain.State.Horses;
-using EMS.Judge.Application.Core;
+using EMS.Judge.Application.Common;
 using System.Collections.Generic;
 using System.Linq;
