+++ src/Judge/EMS.Judge.Application/Queries/ParticipantQueries.cs
@@ -1,6 +1,6 @@
 ﻿using Core.Domain.State;
 using Core.Domain.State.Participants;
-using EMS.Judge.Application.Core;
+using EMS.Judge.Application.Common;
 using System.Collections.Generic;
 using System.Linq;
