+++ src/Judge/EMS.Judge/Common/ValueSerializer.cs
@@ -1,4 +1,4 @@
-﻿using EMS.Judge.Application.Core.Exceptions;
+﻿using EMS.Judge.Application.Common.Exceptions;
 using System;
 using System.Globalization;
 using static EMS.Judge.DesktopConstants;
