+++ src/Judge/EMS.Judge/Views/Content/Configuration/Children/Competitions/AddParticipants/AddParticipantsViewModel.cs
@@ -3,8 +3,8 @@ using EMS.Judge.Common.Services;
 using EMS.Judge.Common.ViewModels;
 using EMS.Judge.Services;
 using EMS.Judge.Application.Services;
-using EMS.Judge.Application.Core;
-using EMS.Judge.Application.Core.Models;
+using EMS.Judge.Application.Common;
+using EMS.Judge.Application.Common.Models;
 using Core.Mappings;
 using Core.Domain.AggregateRoots.Configuration;
 using Core.Domain.State.Participants;
