+++ src/Judge/EMS.Judge/Views/Content/Configuration/Children/Laps/LapViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using EMS.Judge.Services;
 using EMS.Judge.Views.Content.Configuration.Core;
-using EMS.Judge.Application.Core;
+using EMS.Judge.Application.Common;
 using Core.Domain.AggregateRoots.Configuration;
 using Core.Domain.Core.Models;
 using Core.Domain.State.Laps;
