+++ src/Judge/EMS.Judge/Views/Content/Configuration/Children/Personnel/PersonnelViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using EMS.Judge.Common.Components.Templates.SimpleListItem;
 using EMS.Judge.Services;
 using EMS.Judge.Views.Content.Configuration.Core;
-using EMS.Judge.Application.Core;
+using EMS.Judge.Application.Common;
 using Core.Domain.AggregateRoots.Configuration;
 using Core.Domain.Core.Models;
 using Core.Domain.Enums;
