+++ src/Judge/EMS.Judge/Views/Content/Configuration/Roots/Events/EnduranceEventViewModel.cs
@@ -4,7 +4,7 @@ using EMS.Judge.Services;
 using EMS.Judge.Views.Content.Configuration.Children.Competitions;
 using EMS.Judge.Views.Content.Configuration.Children.Personnel;
 using EMS.Judge.Views.Content.Configuration.Core;
-using EMS.Judge.Application.Core;
+using EMS.Judge.Application.Common;
 using EMS.Judge.Application.Queries;
 using Core.Mappings;
 using Core.Domain.AggregateRoots.Configuration;
