+++ src/Judge/EMS.Judge/Views/Content/Configuration/Roots/Events/EnduranceEventViewModelMaps.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using EMS.Judge.Views.Content.Configuration.Children.Personnel;
 using EMS.Judge.Application.Aggregates.Configurations;
-using EMS.Judge.Application.Core.Models;
+using EMS.Judge.Application.Common.Models;
 using Core.Mappings;
 using Core.Domain.State.Countries;
 using Core.Domain.State.EnduranceEvents;
