+++ src/Judge/EMS.Judge/Views/Content/Configuration/Roots/Horses/HorseViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using EMS.Judge.Services;
 using EMS.Judge.Views.Content.Configuration.Core;
-using EMS.Judge.Application.Core;
+using EMS.Judge.Application.Common;
 using Core.Models;
 using Core.Domain.AggregateRoots.Configuration;
 using Core.Domain.Core.Models;
