+++ src/Judge/EMS.Judge/Views/Content/Configuration/Roots/Participants/Listing/ParticipantListViewModel.cs
@@ -2,8 +2,8 @@
 using EMS.Judge.Common.ViewModels;
 using EMS.Judge.Services;
 using EMS.Judge.Application.Services;
-using EMS.Judge.Application.Core;
-using EMS.Judge.Application.Core.Models;
+using EMS.Judge.Application.Common;
+using EMS.Judge.Application.Common.Models;
 using Core.Mappings;
 using Core.Domain.AggregateRoots.Configuration;
 using Core.Domain.State.Participants;
