+++ src/Judge/EMS.Judge/Views/Content/Manager/ParticipantsList/Config/ParticipantMaps.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using EMS.Judge.Application.Core.Models;
+using EMS.Judge.Application.Common.Models;
 using Core.Mappings;
 using Core.Domain.State.Participants;
